refactor: reorganize tests according to nest best practice
