+++ package.json
@@ -21,7 +21,7 @@
     "test:watch": "test jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "dotenv -e .env.test -- jest --runInBand --config ./test/jest-e2e.json",
+    "test:e2e": "dotenv -e .env.test -- jest --runInBand --config ./test/e2e/jest-e2e.json",
     "db": "docker compose up",
     "db:deploy": "docker compose up -d",
     "db:remove": "docker compose down ",
