+++ src/clients/mdw-ws-client.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, Logger } from '@nestjs/common';
 import * as WebSocket from 'ws';
-import NETWORKS from '../lib/networks';
+import NETWORKS from '../lib/network-config';
 import { nonNullable, pluralize } from '../lib/utils';
 import { SubscriptionEvent } from './mdw-ws-client.model';
 import { ContractAddress } from './sdk-client.model';
