+++ src/lib/network-config.ts
