+++ test/test-utils/utils.ts
@@ -1,0 +1,2 @@
+export const sortByAddress = (xs: { address: string }[]) =>
+  [...xs].sort((a, b) => a.address.localeCompare(b.address));
